requirements.txt
PyQt5==5.15.4
keyboard>=0.13.5
PyQtWebEngine>=5.15.4
rsa>=4.7.2
sip>=6.1.1
pycryptodome>=3.10.1